% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_product.R
\name{get_product}
\alias{get_product}
\title{Get product from data file}
\usage{
get_product(cat)
}
\arguments{
\item{cat}{The product category}
}
\value{
A product from the data file
}
\description{
Picks a random product item from the data file based on the category input
}
