% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_custom_tag.R
\name{make_custom_tag}
\alias{make_custom_tag}
\title{Create a product tag for a custom Fakea product}
\usage{
make_custom_tag(
  logo = 2,
  currency = "$",
  language = "sw",
  txt1 = "FEJKA",
  txt2 = "R package",
  txt3 = " v1.0.0",
  destfile = "./fakea_CUSTOM.png"
)
}
\arguments{
\item{logo}{Which type of tag to output (1=blank; 2=logo; 3=Swedish pretty logo)}

\item{currency}{The currency symbol to use for price tag (default="$")}

\item{language}{The language used for product name and logo (default="sw")}

\item{txt1}{The custom product name}

\item{txt2}{The custom product category}

\item{txt3}{The custom price}

\item{destfile}{The destination output file}
}
\value{
A custom product tag image
}
\description{
Outputs a Fakea product tag based on custom inputs
}
