% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tag.R
\name{make_tag}
\alias{make_tag}
\title{Create a product tag for a random Fakea product}
\usage{
make_tag(cat, logo = 2, currency = "$", language = "sw", directory = "./")
}
\arguments{
\item{cat}{The product category}

\item{logo}{Which type of tag to output (1=blank; 2=logo; 3=pretty logo)}

\item{currency}{The currency symbol to use for price tag (default="$")}

\item{language}{The language used for product name and logo (default="sw")}

\item{directory}{The specified directory in which to save your file}
}
\value{
A product tag image
}
\description{
Inputs a product category and outputs a Fakea product tag
}
