% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makea.R
\name{makea}
\alias{makea}
\title{Generate an "IKEA name"}
\usage{
makea(string, original = FALSE, swedeness = 0)
}
\arguments{
\item{string}{The string to be converted}

\item{original}{Whether to use the original or adapted IKEA name rules (default=FALSE)}

\item{swedeness}{How much (0 to 4) the input string should "Swedify" its form (default=0)}
}
\value{
A product from the data file
}
\description{
Inputs a string and outputs the IKEA-fied version (original or adapted rules)
}
