% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swedify.R
\name{swedify}
\alias{swedify}
\title{Swedify an input string}
\usage{
swedify(string, swedeness = 0, invert = FALSE, capitalize = FALSE)
}
\arguments{
\item{string}{The string to be converted}

\item{swedeness}{How much (0 to 2) the input string should "Swedify" its form (default=0)}

\item{invert}{Whether to invert specific di- and trigraphs}

\item{capitalize}{Capitalize first letter of output string}
}
\value{
A product from the data file
}
\description{
Inputs a string and outputs a "Swedified" version
}
